refactor(backfill): split main()-rollback tests out of the CLI test file
The write_failures_occurred fix pushed three files (_backfill_ac_provenance_apply.py,
backfill_ac_provenance.py, test_backfill_ac_provenance_cli.py) over the repo's
300-LOC bloat-gate cap.

Splits main()'s three snapshot/restore tests (orphan tag, mid-batch OSError,
silent write-failure) into a new test_backfill_ac_provenance_cli_rollback.py
sibling, same precedent as that file's own _git.py/_apply_skips.py splits, and
tightens two in-code comments in the production files. No behavior change.
