+++ .shipwright/planning/adr/iterate-2026-09-07-p3-4-tagging-backfill-mechanical-ac-provenance-backfill.md
@@ -131,7 +131,7 @@ turn, T5 disputed with a reproduction rather than fixed)
 | T4 | high | The writer follows repository-controlled paths (`project_root / rel`) and can write through a symlink (or Windows reparse point) outside the project root when run with `--write` — `Path.is_file()` FOLLOWS the final symlink, so it alone never catches this | accepted-and-fixed — added `backfill_write.is_contained(project_root, abs_path)`: rejects a symlinked leaf (`is_symlink()`) AND checks the resolved path is still under the resolved project root (`resolve(strict=True)`, catching a symlinked ancestor DIRECTORY too, not just the leaf). Wired into both call sites that read/write a candidate file before this fix had any containment check at all — `backfill_write.apply_writes` (the shared engine `backfill_test_links.py` also calls, with no containment check of its own) and this tool's own `apply_upgrades` read/write loop. Regression test: a symlinked test file pointing outside the project root is skipped as `path_escapes_project_root` and the real external target is confirmed byte-unmodified, in both call sites |
 | T5 | high (disputed — false positive) | `_upgrade_bare_tags` allegedly replaces only the OPENING quote of a single-quoted `covers('FR-01.01')` bare tag, leaving the original closing quote behind and emitting invalid Python (`covers("FR-01.01/AC01')`) | disputed-and-not-fixed — reproduced directly: `_covers_pattern`'s regex captures the opening quote AND backreferences it (`\1`) to require the SAME character as the closing delimiter, so the matched span always spans the ENTIRE quoted literal (both delimiters), not just the opening one; the hardcoded `covers("{fr_id}/{ac_id}"` replacement therefore replaces the whole literal in one span and always emits valid, consistently double-quoted syntax. Verified by directly invoking `_upgrade_bare_tags` on a single-quoted fixture and asserting `ast.parse` on the rewritten file does not raise — it does not. Added the requested regression test anyway (proves the gate wrong going forward); no production code changed, since none is warranted for a claim that does not reproduce |
 | T6 | high | `apply_upgrades` writes each bare-tag upgrade immediately (`abs_path.write_bytes(...)`) while iterating candidates, uncaught — an I/O failure partway through a multi-file batch (disk full, permission error, a file made read-only mid-run) previously propagated straight out of `main()`, so the existing pre-write snapshot/restore (D3's own safety net, built for the ORPHAN-tag case) never ran and every file already written before the failure stayed modified with no recovery | accepted-and-fixed — `main()`'s call to `apply_upgrades` is now wrapped in `try/except OSError`, restoring every file in the same pre-write snapshot the orphan-rollback path already uses and returning `rc=1` with `{"write_error": ..., "rolled_back": True}`. Regression test: a two-file batch where the SECOND file's write raises `OSError` — the FIRST file (already written) is confirmed restored to its original content, not left half-applied. Advisory (non-blocking) follow-up from the same review round, tracked rather than fixed: multiline/commented `@pytest.mark.covers(...)` decorators are invisible to the current single-line-regex ownership check (`_DECORATOR_LINE_RE`) — a real but pre-existing conservatism (such a tag is silently treated as "no bare tag here" rather than mis-parsed), not a regression this fix introduces; reworking the ownership check to a full AST/decorator-span parse is scope creep against a one-shot backfill unit (`trg-ce51177e`'s sibling — same "next unit that touches this module" disposition as D2) |
-| T7 | high | `apply_writes` can return a `write_failures` entry for an insertion candidate (e.g. a TOCTOU race — `apply_upgrades` already read the file successfully once itself before adding it to the insert batch, `apply_writes` re-reads it a second time at the end) AFTER a SIBLING candidate's bare-tag upgrade elsewhere in the same batch already succeeded and was written directly to disk; `main()` folded `write_failures` into the ordinary `skipped` list and still returned `rc=0` — a partial application reported as an unqualified success, with the sibling upgrade never rolled back | accepted-and-fixed, narrowly scoped — `apply_upgrades` now also returns `write_failures_occurred: bool`, kept DISTINCT from the pre-existing `skipped` reasons (`file_absent_at_head`, `ambiguous_*`, etc. — deliberate, by-design non-writes, never treated as failures; rolling those back too would regress the tool's own documented best-effort-per-file design). `main()` now rolls back on `orphans OR write_failures_occurred`, sharing the same pre-write snapshot both branches already used. Regression test: `main()` with a monkeypatched `apply_upgrades` that writes one real bare-tag upgrade to disk and then reports `write_failures_occurred=True` — the already-written file is confirmed restored to its original content |
+| T7 | high | `apply_writes` can return a `write_failures` entry for an insertion candidate (e.g. a TOCTOU race — `apply_upgrades` already read the file successfully once itself before adding it to the insert batch, `apply_writes` re-reads it a second time at the end) AFTER a SIBLING candidate's bare-tag upgrade elsewhere in the same batch already succeeded and was written directly to disk; `main()` folded `write_failures` into the ordinary `skipped` list and still returned `rc=0` — a partial application reported as an unqualified success, with the sibling upgrade never rolled back | accepted-and-fixed, narrowly scoped — `apply_upgrades` now also returns `write_failures_occurred: bool`, kept DISTINCT from the pre-existing `skipped` reasons (`file_absent_at_head`, `ambiguous_*`, etc. — deliberate, by-design non-writes, never treated as failures; rolling those back too would regress the tool's own documented best-effort-per-file design). `main()` now rolls back on `orphans OR write_failures_occurred`, sharing the same pre-write snapshot both branches already used. Regression test: `main()` with a monkeypatched `apply_upgrades` that writes one real bare-tag upgrade to disk and then reports `write_failures_occurred=True` — the already-written file is confirmed restored to its original content. This round's edits crossed the repo's 300-LOC bloat-gate cap on three files; cleared by splitting `main()`'s three snapshot/restore tests (orphan, mid-batch OSError, silent write-failure) out of `test_backfill_ac_provenance_cli.py` into a new `test_backfill_ac_provenance_cli_rollback.py` sibling (same precedent as that file's own `_git.py`/`_apply_skips.py` splits) and tightening two in-code comments, rather than filing an exception |
 
 T1/T2/T3 are mechanism-level fixes to code this unit itself introduced (not
 the frozen shared `backfill_scan.py`/`fr_tag_grammar.py` engines, which have
