+++ shared/scripts/tools/_backfill_ac_provenance_apply.py
@@ -274,10 +274,8 @@ def apply_upgrades(project_root: Path, report: dict) -> dict:
         "skipped": skipped,
         "tags_upgraded_total": sum(a["tags_upgraded"] for a in upgraded),
         "tags_inserted_total": len(inserted),
-        # Distinct from `skipped` (Tier-3 CI-gate re-review, P3.4 high): every
-        # OTHER skip reason is a decision NOT to write; this one means a write
-        # was ATTEMPTED and FAILED -- callers must not report unqualified
-        # success alongside an already-applied sibling upgrade.
+        # Unlike `skipped` (a decision NOT to write), this means a write was
+        # ATTEMPTED and FAILED (Tier-3 CI-gate re-review, P3.4 high).
         "write_failures_occurred": bool(write_failures),
     }
